+++ src/Information/CorrectedMutualInformation.jl
@@ -16,30 +16,30 @@ function _buslje09{A}(aln, alphabet::A, clusters, lambda, apc)
     mi
 end
 
-# """
-# This function takes a MSA or a file and a `Format` as first arguments.
-# Calculates a Z score and a corrected MI/MIp as described on **Busjle et. al. 2009**
-#
-# Argument, type, default value and descriptions:
-#
-# ```
-#   - lambda      Float64   0.05    Low count value
-#   - clustering  Bool      true    Sequence clustering (Hobohm I)
-#   - threshold             62      Percent identity threshold for clustering
-#   - maxgap      Float64   0.5     Maximum fraction of gaps in positions included in calculation
-#   - apc         Bool      true    Use APC correction (MIp)
-#   - usegap      Bool      false   Use gaps on statistics
-#   - samples     Int       100     Number of samples for Z-score
-#   - fixedgaps   Bool      true    Fix gaps positions for the random samples
-# ```
-#
-# This function returns:
-#
-# ```
-#   - Z score
-#   - MI or MIp
-# ```
-# """
+"""
+`buslje09` takes a MSA or a file and a `Format` as first arguments. It calculates a Z score
+and a corrected MI/MIp as described on **Busjle et. al. 2009**.
+
+keyword argument, type, default value and descriptions:
+
+```
+  - lambda      Float64   0.05    Low count value
+  - clustering  Bool      true    Sequence clustering (Hobohm I)
+  - threshold             62      Percent identity threshold for clustering
+  - maxgap      Float64   0.5     Maximum fraction of gaps in positions included in calculation
+  - apc         Bool      true    Use APC correction (MIp)
+  - samples     Int       100     Number of samples for Z-score
+  - fixedgaps   Bool      true    Fix gaps positions for the random samples
+  - alphabet    ResidueAlphabet UngappedAlphabet()  Residue alphabet to be used
+```  
+
+This function returns:
+
+```
+  - Z score
+  - MI or MIp
+```
+"""
 function buslje09(aln::AbstractMatrix{Residue};
                   lambda::Float64 = 0.05,
                   clustering::Bool = true,
@@ -91,29 +91,30 @@ function _BLMI(aln, clusters, alpha, beta, apc, lambda::Float64=0.0)
     mi
 end
 
-# """
-# This function takes a MSA or a file and a `Format` as first arguments.
-# Calculates a Z score (ZBLMI) and a corrected MI/MIp as described on **Busjle et. al. 2009** but using using BLOSUM62 pseudo frequencies instead of a fixed pseudocount.
-#
-# Argument, type, default value and descriptions:
-#
-# ```
-#   - beta        Float64   8.512   β for BLOSUM62 pseudo frequencies
-#   - lambda      Float64   0.0     Low count value
-#   - threshold             62      Percent identity threshold for sequence clustering (Hobohm I)
-#   - maxgap      Float64   0.5     Maximum fraction of gaps in positions included in calculation
-#   - apc         Bool      true    Use APC correction (MIp)
-#   - samples     Int       50      Number of samples for Z-score
-#   - fixedgaps   Bool      true    Fix gaps positions for the random samples
-# ```
-#
-# This function returns:
-#
-# ```
-#   - Z score (ZBLMI)
-#   - MI or MIp using BLOSUM62 pseudo frequencies (BLMI/BLMIp)
-# ```
-# """
+"""
+`BLMI` takes a MSA or a file and a `Format` as first arguments. It calculates a Z score
+(ZBLMI) and a corrected MI/MIp as described on **Busjle et. al. 2009** but using using
+BLOSUM62 pseudo frequencies instead of a fixed pseudocount.
+
+Keyword argument, type, default value and descriptions:
+
+```
+  - beta        Float64   8.512   β for BLOSUM62 pseudo frequencies
+  - lambda      Float64   0.0     Low count value
+  - threshold             62      Percent identity threshold for sequence clustering (Hobohm I)
+  - maxgap      Float64   0.5     Maximum fraction of gaps in positions included in calculation
+  - apc         Bool      true    Use APC correction (MIp)
+  - samples     Int       50      Number of samples for Z-score
+  - fixedgaps   Bool      true    Fix gaps positions for the random samples
+```
+
+This function returns:
+
+```
+  - Z score (ZBLMI)
+  - MI or MIp using BLOSUM62 pseudo frequencies (BLMI/BLMIp)
+```
+"""
 function BLMI(aln::AbstractMatrix{Residue};
               beta::Float64 = 8.512,
               threshold = 62,