+++ src/Information/Information.jl
@@ -1,10 +1,12 @@
 """
 The `Information` module of MIToS defines types and functions useful to calculate
-information measures (e.g. *Mutual Information* (MI) and *Entropy*) over a Multiple Sequence Alignment (MSA).
-This module was designed to count `Residue`s (defined in the `MSA` module)
-in special contingency tables (as fast as possible) and to derive probabilities from this counts.
-Also, includes methods for applying corrections to that tables, e.g. pseudocounts and pseudo frequencies.
-Finally, `Information` allows to use this probabilities and counts to estimate information measures and other frequency based values.
+information measures (e.g. *Mutual Information* (MI) and *Entropy*) over a Multiple
+Sequence Alignment (MSA). This module was designed to count `Residue`s
+(defined in the `MSA` module) in special contingency tables (as fast as possible) and
+to derive probabilities from this counts. Also, includes methods for applying corrections
+to that tables, e.g. pseudocounts and pseudo frequencies. Finally, `Information` allows to
+use this probabilities and counts to estimate information measures and other
+frequency based values.
 
 **Features**