+++ src/Information/Iterations.jl
@@ -28,8 +28,7 @@ _mapfreq_kargs_doc = """
 """
 
 _mappairfreq_kargs_doc = """
-- `usediagonal` (default: `true`): Indicates if the function should be applied to identical element pairs.
-- `diagonalvalue` (default: `0`): Value to fill diagonal elements if `usediagonal` is `false`.
+- `diagonalvalue` (default: `0`): Value to fill diagonal elements if `usediagonal` is `Val{false}`.
 """
 
 # Residues: The output is a Named Vector
@@ -50,6 +49,10 @@ end
 # Map to each column
 
 """
+It efficiently map a function (first argument) that takes a table of `Counts` or
+`Probabilities` (third argument). The table is filled in place with the counts or
+probabilities of each column from the `msa` (second argument).
+
 $_mapfreq_kargs_doc
 """
 function mapcolfreq!{T,A}(f::Function, msa::AbstractMatrix{Residue},
@@ -69,6 +72,10 @@ end
 # Map to each sequence
 
 """
+It efficiently map a function (first argument) that takes a table of `Counts` or
+`Probabilities` (third argument). The table is filled in place with the counts or
+probabilities of each sequence from the `msa` (second argument).
+
 $_mapfreq_kargs_doc
 """
 function mapseqfreq!{T,A}(f::Function, msa::AbstractMatrix{Residue},
@@ -106,12 +113,18 @@ end
 # Map to column pairs
 
 """
+It efficiently map a function (first argument) that takes a table of `Counts` or
+`Probabilities` (third argument). The table is filled in place with the counts or
+probabilities of each pair of columns from the `msa` (second argument).
+The fourth positional argument `usediagonal` indicates if the function should be applied
+to identical element pairs (default to `Val{true}`).
+
 $_mapfreq_kargs_doc
 $_mappairfreq_kargs_doc
 """
 function mapcolpairfreq!{T,A,D}(f::Function, msa::AbstractMatrix{Residue},
                                 table::Union{Probabilities{T,2,A},Counts{T,2,A}},
-                                ::Type{Val{D}} = Val{true};
+                                usediagonal::Type{Val{D}} = Val{true};
                                 diagonalvalue::T = zero(T),
                                 kargs...)
     ncol = ncolumns(msa)
@@ -126,12 +139,18 @@ end
 # Map to sequence pairs
 
 """
+It efficiently map a function (first argument) that takes a table of `Counts` or
+`Probabilities` (third argument). The table is filled in place with the counts or
+probabilities of each pair of sequences from the `msa` (second argument).
+The fourth positional argument `usediagonal` indicates if the function should be applied
+to identical element pairs (default to `Val{true}`).
+
 $_mapfreq_kargs_doc
 $_mappairfreq_kargs_doc
 """
 function mapseqpairfreq!{T,A,D}(f::Function, msa::AbstractMatrix{Residue},
                                 table::Union{Probabilities{T,2,A},Counts{T,2,A}},
-                                ::Type{Val{D}} = Val{true};
+                                usediagonal::Type{Val{D}} = Val{true};
                                 diagonalvalue::T = zero(T),
                                 kargs...)
     sequences = getresiduesequences(msa)
@@ -140,32 +159,6 @@ function mapseqpairfreq!{T,A,D}(f::Function, msa::AbstractMatrix{Residue},
     _mappairfreq!(f, sequences, plm, table, Val{D}; kargs...)
     scores
 end
-
-# TO DO:
-# N >= 3 using using combinations from Combinatorics.jl
-
-# """
-# `Information.mapcolfreq!(aln, [count,] use, [α, β,] measure, [pseudocount,] [weight,] [usediagonal, diagonalvalue])`
-#
-# This function `estimate` a `AbstractMeasure` in columns or pair of columns of a MSA.
-#
-# - `aln` : This argument is mandatory and it should be a `Matrix{Residue}`. Use the function `getresidues` (from the MSA module) over a MSA object to get the needed matrix.
-# - `count` : It should be defined when `use` is a `ResidueProbability` object. It indicates the element type of the counting table.
-# - `use` : This argument is mandatory and indicates the sub-type of `ResidueContingencyTables` used by `estimate` inside the function.
-# If the table has one dimension (`N`=`1`), the occurrences/probabilities are counted for each sequence/column.
-# If the table has two dimension (`N`=`2`), pairs of sequences/columns are used.
-# The dimension `N` and the `UseGap` parameter of `Residueount{T, N, UseGap}` or `ResidueProbability{T, N, UseGap}` determines the output and behaviour of this functions.
-# If `UseGap` is true, gaps are used in the estimations.
-# - `α` : This argument is optional, and indicates the weight of real frequencies to apply BLOSUM62 based pseudo frequencies.
-# - `β` : This argument is optional, and indicates the weight of BLOSUM62 based pseudo frequencies.
-# - `measure` : This argument is mandatory and indicates the measure to be used by `estimate` inside the function.
-# - `pseudocount` : This argument is optional. It should be an `AdditiveSmoothing` instance (default to zero).
-# - `weight` : This argument is optional. It should be an instance of `ClusteringResult` or `AbstractVector` (vector of weights).
-# Each sequence has weight 1 (`NoClustering()`) by default.
-# - `usediagonal` : This functions return a `Vector` in the one dimensional case, or a `PairwiseListMatrix` in the bidimensional case.
-# This argument only have sense in the bidimensional case and indicates if the list on the `PairwiseListMatrix` should include the diagonal (default to `true`).
-# - `diagonalvalue` : This argument is optional (default to zero). Indicates the value of output diagonal elements.
-# """
 
 # cMI
 # ===